move cmp into osutil (thanks Gustavo)
