+++ policy/policy.go
@@ -26,7 +26,6 @@ import (
 	"os"
 	"path/filepath"
 
-	"github.com/ubuntu-core/snappy/cmp"
 	"github.com/ubuntu-core/snappy/osutil"
 )
 
@@ -131,7 +130,7 @@ func Remove(pkgName, instPath, rootDir string) error {
 }
 
 func aaUp(old, new, dir, pfx string) map[string]bool {
-	return cmp.DirUpdated(filepath.Join(old, dir), filepath.Join(new, dir), pfx)
+	return osutil.DirUpdated(filepath.Join(old, dir), filepath.Join(new, dir), pfx)
 }
 
 // AppArmorDelta returns which policies and templates are updated in the package
